fix command #2430
