+++ src/DataFixtures/Loader/LoadFileData.php
@@ -3,6 +3,7 @@
 namespace App\DataFixtures\Loader;
 
 use App\Entity\Enum\DocumentType;
+use App\Entity\File;
 use App\Factory\FileFactory;
 use App\Repository\UserRepository;
 use Doctrine\Bundle\FixturesBundle\Fixture;
@@ -12,15 +13,6 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
 class LoadFileData extends Fixture implements OrderedFixtureInterface
 {
-    public const STANDALONE_FILES = [
-        '1 - Demande de transmission d\'une copie d\'un DPE' => '1_Demande_de_transmission_d_une_copie_d_un_DPE.docx',
-        '2 - Information au bailleur - Mise en conformité' => '2_Information_au_bailleur_Mise_en_conformite.docx',
-        '3 - Mise en demeure' => '3_Mise_en_demeure.docx',
-        '4 - Invitation à contacter l\'ADIL' => '4_Invitation_a_contacter_l_ADIL.docx',
-        '5 - Engagement du bailleur à réaliser des travaux' => '5_Engagement_du_bailleur_a_realiser_des_travaux.docx',
-        '6 - Saisine de la Commission départementale de conciliation' => '6_Saisine_de_la_Commission_departementale_de_conciliation.docx',
-    ];
-
     public function __construct(
         private readonly ParameterBagInterface $parameterBag,
         private readonly FileFactory $fileFactory,
@@ -39,7 +31,7 @@ class LoadFileData extends Fixture implements OrderedFixtureInterface
         $manager->persist($file);
 
         $userAdmin = $this->userRepository->findOneBy(['email' => $this->parameterBag->get('admin_email')]);
-        foreach (self::STANDALONE_FILES as $title => $filename) {
+        foreach (File::STANDALONE_FILES as $title => $filename) {
             $file = $this->fileFactory->createInstanceFrom(
                 filename: $filename,
                 title: $title,
